+++ ddtrace/contrib/aioredis/patch.py
@@ -19,8 +19,8 @@ from ...ext import db
 from ...ext import net
 from ...ext import redis as redisx
 from ...internal.utils.formats import stringify_cache_args
-from ..redis.util import _trace_redis_cmd
-from ..redis.util import _trace_redis_execute_pipeline
+from ..trace_utils_redis import _trace_redis_cmd
+from ..trace_utils_redis import _trace_redis_execute_pipeline
 
 
 try:
