+++ ddtrace/contrib/aredis/patch.py
@@ -6,8 +6,8 @@ from ddtrace.vendor import wrapt
 from ...internal.utils.formats import stringify_cache_args
 from ...internal.utils.wrappers import unwrap
 from ...pin import Pin
-from ..redis.util import _trace_redis_cmd
-from ..redis.util import _trace_redis_execute_pipeline
+from ..trace_utils_redis import _trace_redis_cmd
+from ..trace_utils_redis import _trace_redis_execute_pipeline
 
 
 config._add("aredis", dict(_default_service="redis"))
