+++ ddtrace/contrib/redis/asyncio_patch.py
@@ -2,8 +2,8 @@ from ddtrace import config
 
 from ...internal.utils.formats import stringify_cache_args
 from ...pin import Pin
-from .util import _trace_redis_cmd
-from .util import _trace_redis_execute_pipeline
+from ..trace_utils_redis import _trace_redis_cmd
+from ..trace_utils_redis import _trace_redis_execute_pipeline
 
 
 #
